invoice crud

+++ database/migrations/2022_02_05_161731_create_bases_table.php
@@ -17,7 +17,7 @@ class CreateBasesTable extends Migration
             $table->id();
             $table->integer('lens_id') ;
             $table->integer('base') ;
-            $table->float('true_base') ;
+            $table->decimal('true_base') ;
             $table->timestamps();
         });
     }
